refactor : 각각의 모드에서 다른 Config를 사용하도록 변경
