+++ detector/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,1 @@
-io.jeyong.detector.config.NPlusOneDetectorConfig
+io.jeyong.detector.config.NPlusOneDetectorLoggingConfig
